+++ codex-rs/mcp-client/Cargo.toml
@@ -11,11 +11,11 @@ serde_json = "1"
 tracing = { version = "0.1.41", features = ["log"] }
 tracing-subscriber = { version = "0.3", features = ["fmt", "env-filter"] }
 tokio = { version = "1", features = [
-    "io-std",
+    "io-util",
     "macros",
     "process",
     "rt-multi-thread",
-    "signal",
+    "sync",
 ] }
 
 [dev-dependencies]
